mv: add advise_to_reapply hint for moving file into cone
Originally, the SKIP_WORKTREE bit is not removed when moving an out-of-cone
file into sparse cone, thus the moved file does not show up in the working tree.
Hint the user to use "git sparse-checkout reapply" to reapply the sparsity rules
to the working tree, by which the SKIP_WORKTREE bit is removed.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>