mv: check if <destination> is a directory in the index with
Originally, <destination> is assumed to be in the working tree. If it is
not found as a directory, then it is determined to be either a regular file
path, or error out if used under the first form of 'git-mv'.

Change the logic so that when supplied with --sparse, first check if
<destination> is a directory with all its contents sparsified (called a
SKIP_WORKTREE_DIR). If yes, then treat <destination> as a directory
exists in the working tree. If no, continue the original checking logic.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>