pathspec.h: move pathspec_needs_expanded_index() from reset.c to here
pathspec_needs_expanded_index() is reusable when we need to verify if
the index needs to be expanded when the command is utilizing a pathspec
rather than a literal path. Move it for reusability.

Add a few items to the function so it can better serve its purpose as
a standalone public function:

* Add a check in front so if the index is not sparse, return early since
  no expansion is needed.

* Add documentation to the function.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>