rm: expand the index only when necessary
Originally, rm a pathspec that is out-of-cone in a sparse-index
environment, Git dies with "pathspec '<x>' did not match any files",
mainly because it does not expand the index so nothing is matched.

Expand the index when the pathspec needs an expanded index, i.e. the
pathspec contains wildcard that may need a full-index or the pathspec
is simply out-of-cone.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>