t1092: normalize a behavioral difference of `git-rm` under sparse-index
`git-rm` a sparse-directory entry within a sparse-index enabled repo
behaves differently from a sparse directory within a sparse-checkout
enabled repo.

For example, in a sparse-index repo, where 'folder1' is a
sparse-directory entry, `git rm -r --sparse folder1` provides this:

	rm 'folder1/'

Whereas in a sparse-checkout repo *without* sparse-index, doing so
provides this:

	rm 'folder1/0/0/0'
	rm 'folder1/0/1'
	rm 'folder1/a'

Because `git rm` a sparse-directory entry does not need to expand the
index, therefore we should accept the current behavior, which is faster
than "expand the sparse-directory entry to match the sparse-checkout
situation".

Modify a previous test so such difference is not considered as an error.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>