builtin/grep.c: walking tree instead of expanding the index
Before this patch, `grep` utilizes the ensure_full_index() method
when --sparse is used, to expand the index and make sure the
content within sparse directories are searched as if they are
regular cache entries in the index.

Now, use grep_tree() instead, to achieve the same result and reduce
overheads from constructing index.

Signed-off-by: Shaoxuan Yuan <shaoxuan.yuan02@gmail.com>